refactor: consolidate jest-dom type declarations into single file
