--- apps/web/src/types/jest-globals-augment.d.ts
@@ -1,30 +1,0 @@
-import '@testing-library/jest-dom';
-
-declare module '@jest/globals' {
-  interface Matchers<R = void> {
-    // @testing-library/jest-dom matchers
-    toBeInTheDocument(): R;
-    toBeVisible(): R;
-    toBeEmpty(): R;
-    toBeDisabled(): R;
-    toBeEnabled(): R;
-    toBeInvalid(): R;
-    toBeRequired(): R;
-    toBeValid(): R;
-    toContainElement(element: Element | null): R;
-    toContainHTML(html: string): R;
-    toHaveAccessibleDescription(text?: string | RegExp): R;
-    toHaveAccessibleName(text?: string | RegExp): R;
-    toHaveAttribute(attr: string, value?: string | number | boolean): R;
-    toHaveClass(...classNames: string[]): R;
-    toHaveFocus(): R;
-    toHaveFormValues(values: Record<string, unknown>): R;
-    toHaveStyle(css: string | Record<string, unknown>): R;
-    toHaveTextContent(text: string | RegExp, options?: { normalizeWhitespace: boolean }): R;
-    toHaveValue(value: string | string[] | number): R;
-    toHaveDisplayValue(value: string | RegExp | (string | RegExp)[]): R;
-    toBeChecked(): R;
-    toBePartiallyChecked(): R;
-    toHaveErrorMessage(text?: string | RegExp): R;
-  }
-}
